♻️ refactor: model계층 pakage 구조 세분화 request, response 구분
